Add pid field to dblp-schema
This is not reported in the dblp, however, it is indeed a field that is
being returned from the API and so it needed to be added.

Overwrite dblp options instead of using only them

the charkey '_value' seems to be necessary for the program to work, so,
the default (dblp.options) was changed to a external xml2jsOptions and
used as a base, overriding the specified fields in parseOptions with the
ones from the default configuration ({...xml2jsOptions, ...parseOptions}).
It had to be taken out of the class since the method is static and
static properties in classes aren't fully supported yet.